refactor: docker-compose dependency support
