+++ pkg/devspace/compose/testdata/basic/devspace.yaml
