+++ pkg/devspace/compose/testdata/build_entry_point/devspace.yaml
