+++ pkg/devspace/compose/testdata/build_network/devspace.yaml
