+++ pkg/devspace/compose/testdata/command/devspace.yaml
