+++ pkg/devspace/compose/testdata/depends_on/docker-compose.yaml
