+++ pkg/devspace/compose/testdata/entry_point/devspace.yaml
