+++ pkg/devspace/compose/testdata/restart-no/devspace.yaml
