+++ pkg/devspace/compose/testdata/restart-on-failure/devspace.yaml
