+++ pkg/devspace/compose/testdata/restart-unless-stopped/devspace.yaml
